+++ smartcontract/programs/doublezero-serviceability/src/state/location.rs
@@ -1,4 +1,8 @@
-use crate::{seeds::SEED_LOCATION, state::accounttype::*};
+use crate::{
+    error::{DoubleZeroError, Validate},
+    seeds::SEED_LOCATION,
+    state::accounttype::*,
+};
 use borsh::{BorshDeserialize, BorshSerialize};
 use solana_program::{account_info::AccountInfo, program_error::ProgramError, pubkey::Pubkey};
 use std::fmt;
@@ -136,6 +140,32 @@ impl TryFrom<&AccountInfo<'_>> for Location {
     }
 }
 
+impl Validate for Location {
+    fn validate(&self) -> Result<(), DoubleZeroError> {
+        // Account type must be Location
+        if self.account_type != AccountType::Location {
+            return Err(DoubleZeroError::InvalidAccountType);
+        }
+        if self.code.len() > 32 {
+            return Err(DoubleZeroError::CodeTooLong);
+        }
+        if self.name.len() > 64 {
+            return Err(DoubleZeroError::NameTooLong);
+        }
+        if self.country.len() != 2 {
+            return Err(DoubleZeroError::InvalidCountryCode);
+        }
+        if self.lat < -90.0 || self.lat > 90.0 {
+            return Err(DoubleZeroError::InvalidLatitude);
+        }
+        if self.lng < -180.0 || self.lng > 180.0 {
+            return Err(DoubleZeroError::InvalidLongitude);
+        }
+
+        Ok(())
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -148,17 +178,19 @@ mod tests {
             index: 123,
             bump_seed: 1,
             reference_count: 0,
-            lat: 123.45,
-            lng: 345.678,
+            lat: 50.45,
+            lng: 50.678,
             loc_id: 1212121,
             code: "test-321".to_string(),
             name: "test-test-test".to_string(),
             country: "US".to_string(),
             status: LocationStatus::Activated,
         };
-
         let data = borsh::to_vec(&val).unwrap();
         let val2 = Location::try_from(&data[..]).unwrap();
+
+        val.validate().unwrap();
+        val2.validate().unwrap();
 
         assert_eq!(val.size(), val2.size());
         assert_eq!(val.owner, val2.owner);
@@ -166,5 +198,147 @@ mod tests {
         assert_eq!(val.lat, val2.lat);
         assert_eq!(val.lng, val2.lng);
         assert_eq!(data.len(), val.size(), "Invalid Size");
+    }
+
+    #[test]
+    fn test_state_location_validate_error_invalid_account_type() {
+        let val = Location {
+            account_type: AccountType::Device, // Should be Location
+            owner: Pubkey::new_unique(),
+            index: 123,
+            bump_seed: 1,
+            reference_count: 0,
+            lat: 10.0,
+            lng: 10.0,
+            loc_id: 1212121,
+            code: "test-321".to_string(),
+            name: "test-test-test".to_string(),
+            country: "US".to_string(),
+            status: LocationStatus::Activated,
+        };
+        let err = val.validate();
+        assert!(err.is_err());
+        assert_eq!(err.unwrap_err(), DoubleZeroError::InvalidAccountType);
+    }
+
+    #[test]
+    fn test_state_location_validate_error_code_too_long() {
+        let val = Location {
+            account_type: AccountType::Location,
+            owner: Pubkey::new_unique(),
+            index: 123,
+            bump_seed: 1,
+            reference_count: 0,
+            lat: 10.0,
+            lng: 10.0,
+            loc_id: 1212121,
+            code: "a".repeat(33), // More than 32
+            name: "test-test-test".to_string(),
+            country: "US".to_string(),
+            status: LocationStatus::Activated,
+        };
+        let err = val.validate();
+        assert!(err.is_err());
+        assert_eq!(err.unwrap_err(), DoubleZeroError::CodeTooLong);
+    }
+
+    #[test]
+    fn test_state_location_validate_error_name_too_long() {
+        let val = Location {
+            account_type: AccountType::Location,
+            owner: Pubkey::new_unique(),
+            index: 123,
+            bump_seed: 1,
+            reference_count: 0,
+            lat: 10.0,
+            lng: 10.0,
+            loc_id: 1212121,
+            code: "test-321".to_string(),
+            name: "a".repeat(65), // More than 64
+            country: "US".to_string(),
+            status: LocationStatus::Activated,
+        };
+        let err = val.validate();
+        assert!(err.is_err());
+        assert_eq!(err.unwrap_err(), DoubleZeroError::NameTooLong);
+    }
+
+    #[test]
+    fn test_state_location_validate_error_invalid_country_code() {
+        let val = Location {
+            account_type: AccountType::Location,
+            owner: Pubkey::new_unique(),
+            index: 123,
+            bump_seed: 1,
+            reference_count: 0,
+            lat: 10.0,
+            lng: 10.0,
+            loc_id: 1212121,
+            code: "test-321".to_string(),
+            name: "test-test-test".to_string(),
+            country: "USA".to_string(), // More than 2 characters
+            status: LocationStatus::Activated,
+        };
+        let err = val.validate();
+        assert!(err.is_err());
+        assert_eq!(err.unwrap_err(), DoubleZeroError::InvalidCountryCode);
+    }
+
+    #[test]
+    fn test_state_location_validate_error_invalid_latitude() {
+        let val_low = Location {
+            account_type: AccountType::Location,
+            owner: Pubkey::new_unique(),
+            index: 123,
+            bump_seed: 1,
+            reference_count: 0,
+            lat: -91.0, // Less than minimum
+            lng: 10.0,
+            loc_id: 1212121,
+            code: "test-321".to_string(),
+            name: "test-test-test".to_string(),
+            country: "US".to_string(),
+            status: LocationStatus::Activated,
+        };
+        let err_low = val_low.validate();
+        assert!(err_low.is_err());
+        assert_eq!(err_low.unwrap_err(), DoubleZeroError::InvalidLatitude);
+
+        let val_high = Location {
+            lat: 91.0, // Greater than maximum
+            ..val_low
+        };
+        let err_high = val_high.validate();
+        assert!(err_high.is_err());
+        assert_eq!(err_high.unwrap_err(), DoubleZeroError::InvalidLatitude);
+    }
+
+    #[test]
+    fn test_state_location_validate_error_invalid_longitude() {
+        let val_low = Location {
+            account_type: AccountType::Location,
+            owner: Pubkey::new_unique(),
+            index: 123,
+            bump_seed: 1,
+            reference_count: 0,
+            lat: 10.0,
+            lng: -181.0, // Less than minimum
+            loc_id: 1212121,
+            code: "test-321".to_string(),
+            name: "test-test-test".to_string(),
+            country: "US".to_string(),
+            status: LocationStatus::Activated,
+        };
+        let err_low = val_low.validate();
+        assert!(err_low.is_err());
+        assert_eq!(err_low.unwrap_err(), DoubleZeroError::InvalidLongitude);
+
+        let val_high = Location {
+            lng: 181.0, // Greater than maximum
+            ..val_low
+        };
+        let err_high = val_high.validate();
+        assert!(err_high.is_err());
+        assert_eq!(err_high.unwrap_err(), DoubleZeroError::InvalidLongitude);
     }
 }