+++ smartcontract/programs/doublezero-serviceability/tests/link_dzx_test.rs
@@ -216,8 +216,8 @@ async fn test_dzx_link() {
         DoubleZeroInstruction::CreateDevice(device::create::DeviceCreateArgs {
             code: "A".to_string(),
             device_type: DeviceType::Switch,
-            public_ip: [10, 0, 0, 1].into(),
-            dz_prefixes: "10.1.0.0/24".parse().unwrap(),
+            public_ip: [100, 0, 0, 1].into(),
+            dz_prefixes: "110.1.0.0/24".parse().unwrap(),
             metrics_publisher_pk: Pubkey::default(),
             mgmt_vrf: "mgmt".to_string(),
         }),
@@ -386,7 +386,7 @@ async fn test_dzx_link() {
         DoubleZeroInstruction::CreateLink(LinkCreateArgs {
             code: "la".to_string(),
             link_type: LinkLinkType::DZX,
-            bandwidth: 100000000,
+            bandwidth: 15_000_000_000,
             mtu: 9000,
             delay_ns: 150000,
             jitter_ns: 5000,
@@ -587,7 +587,7 @@ async fn test_dzx_link() {
             code: Some("la2".to_string()),
             contributor_pk: Some(contributor1_pubkey),
             tunnel_type: Some(LinkLinkType::WAN),
-            bandwidth: Some(2000000000),
+            bandwidth: Some(20_000_000_000),
             mtu: Some(8900),
             delay_ns: Some(15000),
             jitter_ns: Some(5000),
@@ -608,7 +608,7 @@ async fn test_dzx_link() {
         .unwrap();
     assert_eq!(tunnel_la.account_type, AccountType::Link);
     assert_eq!(tunnel_la.code, "la2".to_string());
-    assert_eq!(tunnel_la.bandwidth, 2000000000);
+    assert_eq!(tunnel_la.bandwidth, 20000000000);
     assert_eq!(tunnel_la.mtu, 8900);
     assert_eq!(tunnel_la.delay_ns, 15000);
     assert_eq!(tunnel_la.status, LinkStatus::Activated);
@@ -638,7 +638,7 @@ async fn test_dzx_link() {
         .unwrap();
     assert_eq!(tunnel_la.account_type, AccountType::Link);
     assert_eq!(tunnel_la.code, "la2".to_string());
-    assert_eq!(tunnel_la.bandwidth, 2000000000);
+    assert_eq!(tunnel_la.bandwidth, 20000000000);
     assert_eq!(tunnel_la.mtu, 8900);
     assert_eq!(tunnel_la.delay_ns, 15000);
     assert_eq!(tunnel_la.status, LinkStatus::Deleting);